♻️ Refactor code Exercício 03 - Implemente a Cifra de Vigenère para criptografar e descriptografar uma mensagem
